refactor(activity): move fetchActivityHistory action
