+++ app/reducers/lnd.js
@@ -7,9 +7,7 @@ import { fetchInfo, setHasSynced, infoSelectors } from './info'
 import { putWallet, setActiveWallet, walletSelectors } from './wallet'
 import { onboardingFinished, setSeed } from './onboarding'
 import { fetchChannels } from './channels'
-import { fetchTransactions } from './transaction'
-import { fetchPayments } from './payment'
-import { fetchInvoices } from './invoice'
+import { fetchActivityHistory } from './activity'
 
 // ------------------------------------
 // Constants
@@ -107,20 +105,6 @@ export const lightningGrpcActive = (event, lndConfig) => async dispatch => {
   //fetch user data. channels, balances, payments, invoices and transactions
   dispatch(fetchChannels())
   dispatch(fetchActivityHistory())
-}
-
-/**
- * Fetches user activity history. Which includes:
- * Balance
- * Payments
- * Invoices
- * Transactions
- */
-export const fetchActivityHistory = () => dispatch => {
-  dispatch(fetchBalance())
-  dispatch(fetchPayments())
-  dispatch(fetchInvoices())
-  dispatch(fetchTransactions())
 }
 
 // Connected to WalletUnlocker gRPC interface (lnd is ready to unlock or create wallet)
